Setting Sail
I implemented motion. The method I used is pretty different from the
Euler's method most such games use; instead, I keep a running List of
every time an object changes velocity and use that to calculate its
position at any give time. That will come in handy later when i deal
with the Doppler effect and whatnot.
As a test, I plopped in a battleship cruising along at the speed of
light. That's how fast information will travel in my game.
Also, the earth and moon are now to scale.