Odds and Ends
I added a blip sound effect for when a ship does not have enough energy
to complete an order, to help the player understand why nothing
happened. I also touched up my images so that text is readable when you
hover over a button, and started playing around with putting text in
text-boxes.