A Window to the Past
I finally implemented the mechanic that was why I made this game: delay
of information. Everything you see (besides yourself and the HUD) is now
as it was in the past. The game calculates where the object was when it
reflected the light that is now reaching you. It's not perfect yet, but
it will be. It will be.