Two Sides to Every Space Battle
I implemented a simple 180-rotation to the non-host such that everyone
sees the red ships on the right and the blue ships on the left. This
should make things easier for people when playing multiple games, and
makes it much more difficult to screen-peek.
I also added a ding-type sound effect to the radar and made the ping
sprite brighter.